Copies then heavily modifes processor info and utilization from internal coms project
I have been using a couple of internal projects to wrap my head around HTTP/TCP/TLS communications as well as starting to figure out what can be provided in terms of data for a home lab monitoring.

However, each time, I get to a point where I want to combine both, but the data isn't quite.. compatible. So, for now, I've decided to once again create another project, but this time, not internal, and the data output is JSON for now.

This will then be aligned with the coms work, but that will need integration work as well, and by integration I mean copy + heavy modifications as well. The structure is.. okay, but not something I'm not ready to "publish" via this repo just yet.